[context] add socket connection to group context
